% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Independent.R
\name{Independent}
\alias{Independent}
\title{Independent Beta Prior for Bayesian Basket Trial}
\usage{
Independent(nDat, yDat, be.a0 = NULL, be.b0 = NULL)
}
\arguments{
\item{nDat}{a vector of length B for the sample size in each basket.}

\item{yDat}{a vector of length B for the number of responses in each basket.}

\item{be.a0}{a vector of length B for beta prior parameter a0 in each basket.}

\item{be.b0}{a vector of length B for beta prior parameter b0 in each basket.}
}
\value{
It returns a list including the posterior beta parameters.
}
\description{
Independent Beta Prior for Bayesian Basket Trial
}
\examples{
Independent(nDat = c(25, 25, 25, 25, 25), yDat = c(2,9,11,13,20),
be.a0 = rep(0.5, 5), be.b0 = rep(0.5, 5))
}
